fix bug li
Next up in your addPhraseToDisplay you’re creating an li variable once which you’re appending to the list instead of the newLi that you’re creating for each of the characters. You’ll want to remove the li variable and its references and change phraseList.append("li") to be phraseList.append(newLi) as this is the list item that you created for this character and are adding the space and letter classes to.